django-api/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
semantic-version==2.10.0
setuptools==72.1.0
setuptools-rust==1.9.0
sqlparse==0.5.1
